Fix find expressions that check for occurrence of a string
The find expressions in the line_style module where missing the `> -1`
check. Without it the expression was only truthy if the position of the
substring is > 0. I think the intention is for the find expression
to simply test if the substring occurs at any position in the string
being searched (as is the case in the other style code and description
functions).